Menambahkan routes data pekerjaan

+++ app/Http/Controllers/JobController.php
@@ -140,6 +140,17 @@ class JobController extends Controller
      */
     public function destroy($id)
     {
+
+        $connectedDataJob = DB::table('employees')
+        ->where('job_id', $id)
+        ->exists();
+        
+        if ($connectedDataJob) {
+            return redirect()
+                ->route('jobs.index')
+                ->with('warning', 'Pekerjaan tidak bisa dihapus karena masih ada data yang terhubung! Hapus data yang terkait terlebih dahulu di data karyawan.');
+        }
+
         // $job->delete();
         DB::delete('delete from jobs where job_id = ?', [$id]);
